Introduce Resource class in flame
Currently flame TemplateGenerator generates the heat template while
visiting OpenStack (nova,neutron,cinder) resources. It implies lots
of code duplication associated to the template generation.

This change introduces a Resource class in order to split the heat
template generation in 2 parts:
 * Extraction: extracting OpenStack resources and creating Resource
   instances
 * Generation: generating heat template from Resource instances.

It allows to:
 * Reduce code duplication associated to the template generation,
 * Allow to test smaller methods,
 * Add steps between the extraction and the generation (for
   customization for example).

This change will be followed by daughter changes in order to:
 * Improve tests by taking advantage of Resource use,
 * Improve flameclient.flame by simplifying relationships between heat
   template resources and also with parameters.

Change-Id: If5e26c6210145103e729d42b149f177439acaddb