Add an extract-ports flag
When the extract-ports flag is used, flame extracts the neutron ports
and connects them (according to the other flags) to :
 - floating ips
 - nova:compute instances
 - security groups
 - subnets and networks

This allows the exported stack to have the same network configuration as
its source.

When the flag is used in conjunction with --generate-stack-data,
the generated stack data also contains the assocation between
ports and floating IPs.

Co-Authored-By: zarrouk <mzarrouk.ext@orange.com>

Change-Id: Id2e406c7aac5457cdef822b78edf7e30bd833a2d