Fix DHCP server IP address collision with VM's IP
When a stack is imported in Heat, the DHCP server IP is set
to the lowest free IP address of it's pool.
Depending on the VM creation order, the DHCP address can
either collide with vm1's or vm2's IP.
with the alter-allocation-pools flag the exported
allocation pools starts at the DHCP's IP address.

Co-Authored-By: Guillaume Espanel <guillaume.espanel@objectif-libre.com>
Change-Id: I8a4815e6c8b2bafa539cf3ad8384be76ef075edc